+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+- **Fixed:** Handling of `default` attribute in `JSONSchemaBridge`. [\#569](https://github.com/vazco/uniforms/issues/569)
 - **Fixed:** Rendering placeholder in `SelectField` in `uniforms-antd`. [\#568](https://github.com/vazco/uniforms/issues/568)
 
 ## [v2.3.0](https://github.com/vazco/uniforms/tree/v2.3.0) (2019-08-01)